+++ main/camera.h
@@ -19,8 +19,8 @@
 #endif
 
 // How much image (central area) is displayed on screen
-size_t camera_displayed_image_width(void);
-size_t camera_displayed_image_height(void);
+uint16_t camera_displayed_image_width(void);
+uint16_t camera_displayed_image_height(void);
 
 // Function to process images from the camera.
 // Should return false if processing incomplete (and so should be called again with the next frame)
